+++ tests/KcpSharp.ThroughputBanchmarks/StreamThroughputBenchmark/StreamThroughputBenchmarkClient.cs
@@ -47,9 +47,9 @@ namespace KcpSharp.ThroughputBanchmarks.StreamThroughputBenchmark
                 BufferPool = allocator,
                 Mtu = mtu,
                 SendQueueSize = queueSize,
-                SendWindow = windowSize,
-                RemoteReceiveWindow = windowSize,
-                UpdateInterval = updateInterval,
+                SendWindow = (uint)windowSize,
+                RemoteReceiveWindow = (uint)windowSize,
+                UpdateInterval = (uint)updateInterval,
                 NoDelay = noDelay,
                 StreamMode = true
             };
