+++ tests/KcpSharp.ThroughputBanchmarks/StreamThroughputBenchmark/StreamThroughputBenchmarkServer.cs
@@ -32,9 +32,9 @@ namespace KcpSharp.ThroughputBanchmarks.StreamThroughputBenchmark
             {
                 BufferPool = allocator,
                 Mtu = mtu,
-                SendWindow = windowSize,
-                RemoteReceiveWindow = windowSize,
-                UpdateInterval = updateInterval,
+                SendWindow = (uint)windowSize,
+                RemoteReceiveWindow = (uint)windowSize,
+                UpdateInterval = (uint)updateInterval,
                 NoDelay = noDelay,
                 StreamMode = true
             };
